fix: Race condition causing grabFreeJob to not mutated the Coordinator
This resulted in 2000+- less words and this also caused all workers to
grab the same next free job despite locks. The reason those words were
missing was cause all 3 workers would mark the file mapID as done, but
the first worker would change it to be in Reducing status. And the mapID
and reduceID values overlap. So reduceID 7 would be marked as done
without actually being done.